+++ backend/fourletterdisplay.py
@@ -2,45 +2,53 @@
 # -*- coding: utf-8 -*-
 
 import importlib
+import sys
+from datetime import datetime
 from cleep.core import CleepRenderer
 from cleep.common import CATEGORIES
 from cleep.profiles.messageprofile import MessageProfile
-from .fourLetterPHatDriver import FourLetterPHatDriver
+from cleep.profiles.alarmprofile import AlarmProfile
+from .fourletterphatdriver import FourLetterPHatDriver
 
-# use for global lib import
+# used for global lib import
 FOUR_LETTER_PHAT = None
+
 
 class Fourletterdisplay(CleepRenderer):
     """
     Fourletterdisplay application
     """
-    MODULE_AUTHOR = 'Cleep'
-    MODULE_VERSION = '1.0.3'
+
+    MODULE_AUTHOR = "Cleep"
+    MODULE_VERSION = "1.1.0"
     MODULE_DEPS = []
-    MODULE_DESCRIPTION = 'Four-letter pHAT display'
-    MODULE_LONGDESCRIPTION = ('This application installs all needed to use Four-letter pHAT from Piromoni.<br>'
-                             'It allows to display informations from your device (like weather temperature, CPU usage and more) '
-                             'through four letters.<br>'
-                             'Long text (more than 4 letters) will scroll automatically and displayed dots can be '
-                             'configured individually.<br>'
-                             'Brightness is also configurable with a night mode with a night mode')
-    MODULE_TAGS = ['display', 'pimoroni', 'digits', 'segments']
+    MODULE_DESCRIPTION = "Four-letter pHAT display"
+    MODULE_LONGDESCRIPTION = (
+        "This application installs all needed to use Four-letter pHAT from Piromoni.<br>"
+        "It allows to display informations from your device (like weather temperature, CPU usage and more) "
+        "through four letters.<br>"
+        "Long text (more than 4 letters) will scroll automatically and displayed dots can be "
+        "configured individually.<br>"
+        "Brightness is also configurable with a night mode with a night mode"
+    )
+    MODULE_TAGS = ["display", "pimoroni", "digits", "segments"]
     MODULE_CATEGORY = CATEGORIES.DRIVER
     MODULE_COUNTRY = None
-    MODULE_URLINFO = 'https://github.com/tangb/cleepmod-fourletterhat'
-    MODULE_URLHELP = 'https://github.com/tangb/cleepmod-teleinfo/wiki'
-    MODULE_URLSITE = 'https://shop.pimoroni.com/products/four-letter-phat'
-    MODULE_URLBUGS = 'https://github.com/tangb/cleepmod-teleinfo/issues'
+    MODULE_URLINFO = "https://github.com/CleepDevice/cleepmod-fourletterhat"
+    MODULE_URLHELP = "https://github.com/CleepDevice/cleepmod-teleinfo/wiki"
+    MODULE_URLSITE = "https://shop.pimoroni.com/products/four-letter-phat"
+    MODULE_URLBUGS = "https://github.com/CleepDevice/cleepmod-teleinfo/issues"
 
-    MODULE_CONFIG_FILE = 'fourletterpdisplay.conf'
+    MODULE_CONFIG_FILE = "fourletterpdisplay.conf"
     DEFAULT_CONFIG = {
-        'brightness': 15,
-        'nightmode': False,
-        'nightbrightness': 4,
+        "currentbrightness": 15,
+        "brightness": 15,
+        "nightmode": False,
+        "nightbrightness": 4,
     }
 
-    RENDERER_PROFILES = [MessageProfile]
-    RENDERER_TYPE = 'display'
+    RENDERER_PROFILES = [MessageProfile, AlarmProfile]
+    RENDERER_TYPE = "display"
 
     def __init__(self, bootstrap, debug_enabled):
         """
@@ -56,21 +64,26 @@ class Fourletterdisplay(CleepRenderer):
         self.driver = FourLetterPHatDriver()
         self._register_driver(self.driver)
         self.is_night_mode = False
+        self.__enabled_dots = [False, False, False, False]
 
-    def _configure(self):
+    def _on_start(self):
         """
-        Configure module.
+        App started
         """
-        # set configured brightness
         try:
-            self.set_brightness(self._get_config_field('brightness'))
+            self.__change_brightness(self._get_config_field("currentbrightness"))
         except Exception:
             # drop exception when hat is not configured
             pass
 
+        # set current time asap
+        now = datetime.now()
+        time_str = f"{now.hour:02}{now.minute:02}"
+        self.__display_time(time_str)
+
     def _on_stop(self):
         """
-        Stop module
+        Stop app
         """
         try:
             self.clear()
@@ -82,7 +95,7 @@ class Fourletterdisplay(CleepRenderer):
         """
         Event received
 
-        Params:
+        Args:
             event (dict): MessageRequest as dict with event values::
 
                 {
@@ -94,16 +107,22 @@ class Fourletterdisplay(CleepRenderer):
                 }
 
         """
-        if event['event'].endswith('time.sunrise') and self._get_config_field('nightmode'):
-            brightness = self._get_config_field('brightness')
-            self.logger.info('Disable night mode (set brightness to %s/15)' % brightness)
-            self.change_brightness(brightness)
+        if event["event"].endswith("time.sunrise") and self._get_config_field(
+            "nightmode"
+        ):
+            brightness = self._get_config_field("brightness")
+            self.logger.info("Disable night mode (set brightness to %s/15)", brightness)
+            self.__change_brightness(brightness)
             self.is_night_mode = True
 
-        if event['event'].endswith('time.sunset') and self._get_config_field('nightmode'):
-            brightness = self._get_config_field('nightbrightness')
-            self.logger.info('Enable night mode (restore brightness to %s/15)' % brightness)
-            self.change_brightness(brightness)
+        if event["event"].endswith("time.sunset") and self._get_config_field(
+            "nightmode"
+        ):
+            brightness = self._get_config_field("nightbrightness")
+            self.logger.info(
+                "Enable night mode (restore brightness to %s/15)", brightness
+            )
+            self.__change_brightness(brightness)
             self.is_night_mode = False
 
     def on_render(self, profile_name, profile_values):
@@ -111,12 +130,36 @@ class Fourletterdisplay(CleepRenderer):
         Render profile
 
         Args:
-            profile_name (string): rendered profile name
-            values (dict): profile values
+            profile_name (str): rendered profile name
+            profile_values (dict): profile values
         """
-        if profile_name == 'MessageProfile':
-            self.display_message(values['message'])
-            self.set_dots(middle_left=True)
+        if profile_name == "MessageProfile":
+            self.__display_time(profile_values["message"])
+        if profile_name == "AlarmProfile":
+            if profile_values["status"] in (
+                AlarmProfile.STATUS_SCHEDULED,
+                AlarmProfile.STATUS_UNSCHEDULED,
+            ):
+                self.__display_indicator(profile_values["count"] != 0)
+
+    def __display_time(self, time):
+        """
+        Display time (with dot separator)
+
+        Args:
+            time (str): time to display (HHMM)
+        """
+        self.display_message(time)
+        self.set_dots(middle_left=True)
+
+    def __display_indicator(self, turn_on):
+        """
+        Turn on/off indicator (most right LED)
+
+        Args:
+            turn_on (bool): True to turn on indicator, False otherwise
+        """
+        self.set_dots(most_right=turn_on)
 
     def __import_lib(self):
         """
@@ -125,14 +168,18 @@ class Fourletterdisplay(CleepRenderer):
         Raises:
             Exception if driver not installed or lib not installed or screen not connected
         """
+        if "fourletterphat" in sys.modules:
+            self.logger.debug('Module "fourletterphat" is already loaded')
         if not self.driver.is_installed():
-            raise Exception('Four-letter pHAT driver is not installed')
+            raise Exception("Four-letter pHAT driver is not installed")
 
         try:
             global FOUR_LETTER_PHAT
-            FOUR_LETTER_PHAT = importlib.import_module('fourletterphat')
+            FOUR_LETTER_PHAT = importlib.import_module("fourletterphat")
         except Exception as error:
-            raise Exception('Four-letter pHAT does not seem connected. Please check hardware') from error
+            raise Exception(
+                "Four-letter pHAT does not seem connected. Please check hardware"
+            ) from error
 
     def enable_night_mode(self, enable):
         """
@@ -141,39 +188,39 @@ class Fourletterdisplay(CleepRenderer):
         Args:
             enable (bool): Enable night mode
         """
-        self._check_parameters([
-            {'name': 'enable', 'value': enable, 'type': bool}
-        ])
+        self._check_parameters([{"name": "enable", "value": enable, "type": bool}])
 
-        self._set_config_field('nightmode', enable)
+        self._set_config_field("nightmode", enable)
 
-        # restore configured brightness
         if enable and self.is_night_mode:
-            self.set_brightness(self._get_config_field('nightbrightness'))
+            self.__change_brightness(self._get_config_field("nightbrightness"))
         else:
-            self.set_brightness(self._get_config_field('brightness'))
+            self.__change_brightness(self._get_config_field("brightness"))
 
     def set_night_mode_brightness(self, brightness):
         """
         Set nightmode brightness
 
+        Args:
             brightness (int): brighness value (0..15)
         """
-        self._check_parameters([
-            {
-                'name': 'brightness',
-                'value': brightness,
-                'type': int,
-                'validator': lambda val: 0 <= val <= 15,
-                'message': 'Parameter "brightness" must be between 0..15'
-            },
-        ])
+        self._check_parameters(
+            [
+                {
+                    "name": "brightness",
+                    "value": brightness,
+                    "type": int,
+                    "validator": lambda val: 0 <= val <= 15,
+                    "message": 'Parameter "brightness" must be between 0..15',
+                },
+            ]
+        )
 
-        self._set_config_field('nightbrightness', brightness)
+        self._set_config_field("nightbrightness", brightness)
 
         # change brightness
         if self.is_night_mode:
-            self.change_brightness(brightness)
+            self.__change_brightness(brightness)
 
     def clear(self):
         """
@@ -185,17 +232,16 @@ class Fourletterdisplay(CleepRenderer):
 
     def display_message(self, message):
         """
-        Display specified message
+        Display specified message (only 4 chars displayed)
 
         Args:
             message (string): message to display
         """
-        self._check_parameters([
-            {'name': 'message', 'value': message, 'type': str}
-        ])
+        self._check_parameters([{"name": "message", "value": message, "type": str}])
 
         self.__import_lib()
-        FOUR_LETTER_PHAT.scroll_print(message)
+        FOUR_LETTER_PHAT.print_str(message)
+        FOUR_LETTER_PHAT.show()
 
     def set_brightness(self, brightness):
         """
@@ -204,38 +250,68 @@ class Fourletterdisplay(CleepRenderer):
         Args:
             brightness (int): brighness value (0..15)
         """
-        self._check_parameters([
-            {
-                'name': 'brightness',
-                'value': brightness,
-                'type': int,
-                'validator': lambda val: 0 <= val <= 15,
-                'message': 'Parameter "brightness" must be between 0..15',
-            },
-        ])
+        self._check_parameters(
+            [
+                {
+                    "name": "brightness",
+                    "value": brightness,
+                    "type": int,
+                    "validator": lambda val: 0 <= val <= 15,
+                    "message": 'Parameter "brightness" must be between 0..15',
+                },
+            ]
+        )
 
         # save value
-        self._set_config_field('brightness', brightness)
+        self._set_config_field("brightness", brightness)
 
         # change brightness
         if not self.is_night_mode:
-            self.change_brightness(brightness)
+            self.__change_brightness(brightness)
 
-    def change_brightness(self, brightness):
+    def __change_brightness(self, brightness):
         """
         Change brightness
+
+        Args:
+            brightness (int): brighness value (0..15)
         """
         self.__import_lib()
         FOUR_LETTER_PHAT.set_brightness(brightness)
 
-    def set_dots(self, most_left=False, middle_left=False, middle_right=False, most_right=False):
+        # store current brightness to be able to restore it after restart
+        self._set_config_field("currentbrightness", brightness)
+
+    def set_dots(
+        self, most_left=None, middle_left=None, middle_right=None, most_right=None
+    ):
         """
         Turn on/off specified dots
-        """
-        self.__import_lib()
-        FOUR_LETTER_PHAT.set_decimal(0, most_left)
-        FOUR_LETTER_PHAT.set_decimal(1, middle_left)
-        FOUR_LETTER_PHAT.set_decimal(2, middle_right)
-        FOUR_LETTER_PHAT.set_decimal(3, most_right)
-        FOUR_LETTER_PHAT.show()
 
+        Args:
+            most_left (bool): True to turn on, False to turn off, None to let current state
+            middle_left (bool): True to turn on, False to turn off, None to let current state
+            middle_right (bool): True to turn on, False to turn off, None to let current state
+            most_right (bool): True to turn on, False to turn off, None to let current state
+        """
+        self.logger.debug(
+            "set dots: [%s][%s][%s][%s]",
+            most_left,
+            middle_left,
+            middle_right,
+            most_right,
+        )
+        self.__import_lib()
+        if most_left is not None:
+            self.__enabled_dots[0] = most_left
+        if middle_left is not None:
+            self.__enabled_dots[1] = middle_left
+        if middle_right is not None:
+            self.__enabled_dots[2] = middle_right
+        if most_right is not None:
+            self.__enabled_dots[3] = most_right
+
+        for led, value in enumerate(self.__enabled_dots):
+            FOUR_LETTER_PHAT.set_decimal(led, value)
+
+        FOUR_LETTER_PHAT.show()