+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
-## [UNRELEASED]
+## [1.1.0] - 2023-03-13
 ### Changed
 * Update on_render parameter names according to core
+* Do not display scrolling text. Limit text to 4 chars
+* Optimize hardware module loading
+
+### Fixed
+* Restore last brightness after restart (#1388)
+
+### Added
+* Set time at app startup (#1405)
+* Display alarm indicator (#1406)
 
 ## [1.0.3] - 2021-06-04
 ### Changed